pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pr_agent"]

[project]
name = "pr-agent"
version = "0.1.0"
description = "A personal assistant powered by AI. It helps you read, comment, and work with pull requests effortlessly."
dependencies = [
    "anthropic==0.82.0",
    "python-dotenv==1.2.1",
    "pydantic==2.12.5"
]